LeadsDataHub/requirements.txt
amqp==2.6.1
annotated-types==0.6.0
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==21.2.0
authorizenet==1.1.4
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.4
blis==0.7.11
Brlapi==0.8.3
cachetools==5.3.1
captcha==0.4
catalogue==2.0.10
celery==4.4.5
celery-worker==0.0.6
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cms==0.0.2
colorama==0.4.4
command-not-found==0.3
compose==1.5.0
confection==0.1.4
coolname==2.2.0
cron-descriptor==1.4.0
cryptography==35.0.0
cssselect2==0.7.0
cupshelpers==1.0
cymem==2.0.8
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==3.2.13
django-celery==3.1.17
django-celery-beat==2.5.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-classy-tags==3.0.1
django-cms==3.8.2
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filer==3.0.3
django-formtools==2.4
django-js-asset==2.0.0
django-money==2.1
django-mptt==0.15.0
django-object-actions==3.0.2
django-phonenumber-field==5.2.0
django-polymorphic==3.1.0
django-redis==5.3.0
django-sekizai==3.0.1
django-storages==1.13.2
django-timezone-field==6.0.1
django-treebeard==4.4
django-waffle==4.0.0
djangocms-admin-style==2.0.2
djangocms-attributes-field==3.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.1.1
djangocms-icon==2.0.0
djangocms-link==3.1.1
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
docker==5.0.3
dockerpty==0.4.1
docopt==0.6.2
dsdev-utils==1.3.0
duplicity==0.8.21
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
fasteners==0.14.1
fields==5.0.0
filelock==3.12.4
flake8==6.0.0
future==0.18.2
google-api-core==2.11.1
google-auth==2.17.0
google-cloud-core==2.3.3
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==3.0.1
gunicorn==20.0.4
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kombu==4.6.11
kubernetes==26.1.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
Menus==0.2.0
monotonic==1.6
more-itertools==8.10.0
murmurhash==1.0.10
netifaces==0.11.0
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46
openpyxl==3.1.2
packaging==23.2
pandas==2.0.2
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.12.34
Pillow==9.0.1
platformdirs==3.10.0
playwright==1.40.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.1
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py-moneyed==1.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
pyflakes==3.0.1
pygame==2.5.2
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-crontab==3.0.0
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2023.3
PyXB==1.2.5
pyxdg==0.27
PyYAML==6.0.1
redis==5.0.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
sentry-sdk==1.29.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
srsly==2.4.8
stripe==5.4.0
svglib==1.5.1
systemd-python==234
texttable==1.6.4
thinc==8.2.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.24.5
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.2
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
